+++ CHANGES.md
@@ -22,6 +22,7 @@ libcups v3.0b1 (Month DD, YYYY)
 - Added a roll to the default color ippeveprinter printer.
 - Added new DNS-SD API (Issue #19)
 - Added new PWG media sizes.
+- Added new `EXPECT-ALL-VALUES-FROM` predicate for ipptool files (Issue #20)
 - Added, modernized, and promoted the localization interfaces to public API
   (Issue #24)
 - Added public JSON API (Issue #31)